+++ jlt/vector.hpp
@@ -82,31 +82,6 @@ vector(std::initializer_list<T> _l) : std::vector<T>(_l) {}
       return std::vector<T>::operator=(v);
     }
 
-  std::ostream& printMathematicaForm(std::ostream& strm,
-				     const std::string name = "",
-				     const std::string description = "") const
-    {
-      if (this->empty()) return strm;
-
-      // Print description as comment if specified.
-      if (!description.empty())
-	strm << "(* " << description << " *)" << std::endl;
-
-      // Only print = if variable name is specified.
-      if (!name.empty()) strm << name << " = ";
-
-      strm << "{";
-      for (auto i = this->cbegin(); i != this->cend()-1; ++i)
-	{
-	  strm << *i << ",";
-	}
-      // Don't append newline, since in Mathematica it is common to
-      // write on same line.
-      strm << this->back() << "}";
-
-      return strm;
-    }
-
 };
 
 } // namespace jlt